+++ packages/@jacco-meijer/content-models/src/components/opener-section/graphql.ts
@@ -10,6 +10,7 @@ export const fragment = `
   fragment OpenerSection on ContentfulOpenerSection {
     __typename
     heading
+    lightHeading
     preHeading
     subHeading
     image {