+++ packages/@jacco-meijer/content-models/src/templates/config/graphql-fragment.ts
@@ -1,8 +1,8 @@
+import { Vars } from '../../types/vars'
 import { componentFields } from '../helpers/component-fields'
 import { getGraphqlFields } from '../partials/get-graphql-fields'
-import { Vars } from '../types/vars'
 
-export const graphqlFragmentQuery = (vars: Vars) => {
+export const graphqlFragment = (vars: Vars) => {
   // Skip this template if not present in filter array
   if (!vars.options.filter.includes(vars.answers.wheelroomType)) {
     return
@@ -23,21 +23,6 @@ export const fragment = \`
   fragment ${vars.componentName.pascalCase} on Contentful${vars.componentName.pascalCase} {
 ${graphqlFields}
   }
-\`
-
-export const query = \`
-{
-  ${vars.componentName.camelCase}: allContentful${vars.componentName.pascalCase}(
-    limit: 10
-  ) {
-    edges {
-      node {
-        id
-        slug
-      }
-    }
-  }
-}
 \`
 `
 }