+++ packages/@jacco-meijer/content-models/src/templates/partials/get-variation-list.ts
@@ -1,0 +1,11 @@
+import { Vars } from '../../types/vars'
+
+export const getVariationList = (variations: string[], vars: Vars) =>
+  variations
+    .map(
+      (
+        fieldName: string
+      ) => `  '${fieldName}': ${vars.componentName.pascalCase}BasicVar,
+`
+    )
+    .join('')