+++ packages/boilerplate/src/packages/templates/demo/demo.ts
@@ -1,0 +1,8 @@
+import { Vars } from '../types/vars'
+
+export const demo = (vars: Vars) =>
+  `# Demo
+## About
+
+The component name is: ${vars.componentName.camelCase}
+`