+++ packages/boilerplate/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.5](https://github.com/jaccomeijer/wheelroom/compare/boilerplate@2.0.0-beta.4...boilerplate@2.0.0-beta.5) (2019-07-23)
+
+**Note:** Version bump only for package boilerplate
+
+
+
+
+
+# [2.0.0-beta.4](https://github.com/jaccomeijer/wheelroom/compare/boilerplate@2.0.0-beta.3...boilerplate@2.0.0-beta.4) (2019-07-23)
+
+**Note:** Version bump only for package boilerplate
+
+
+
+
+
+# [2.0.0-beta.3](https://github.com/jaccomeijer/wheelroom/compare/boilerplate@2.0.0-beta.2...boilerplate@2.0.0-beta.3) (2019-07-23)
+
+**Note:** Version bump only for package boilerplate
+
+
+
+
+
+# [2.0.0-beta.2](https://github.com/jaccomeijer/wheelroom/compare/boilerplate@2.0.0-beta.1...boilerplate@2.0.0-beta.2) (2019-07-23)
+
+**Note:** Version bump only for package boilerplate
+
+
+
+
+
 # [2.0.0-beta.1](https://github.com/jaccomeijer/wheelroom/compare/boilerplate@2.0.0-beta.0...boilerplate@2.0.0-beta.1) (2019-07-22)
 
 **Note:** Version bump only for package boilerplate